Move TypeScript adapter constructs to new folder
