+++ packages/open-schema-type-script/src/example/ciYamlFile/assertableCiYamlFileContents.ts
@@ -1,9 +1,9 @@
 import fs from 'fs';
 import yaml from 'yaml';
 import { Estinant2 } from '../../core/estinant';
-import { Grition } from '../core/grition';
-import { Odeshin, ODESHIN_GEPP } from '../core/odeshin';
-import { Plifal } from '../core/plifal';
+import { Grition } from '../../type-script-adapter/grition';
+import { Odeshin, ODESHIN_GEPP } from '../../type-script-adapter/odeshin';
+import { Plifal } from '../../type-script-adapter/plifal';
 import {
   ActualCiYamlFilePlifal,
   ACTUAL_CI_YAML_FILE_GEPPP,
