+++ packages/open-schema-type-script/src/example/file/fileA.ts
@@ -8,9 +8,9 @@ import {
 } from '../../utilities/file/fileExtensionSuffixIdentifier';
 import { getFileMetadata } from '../../utilities/file/getFileMetadata';
 import { getNestedFilePaths } from '../../utilities/file/getNestedFilePaths';
-import { Grition } from '../core/grition';
-import { Odeshin, ODESHIN_GEPP } from '../core/odeshin';
-import { Plifal } from '../core/plifal';
+import { Grition } from '../../type-script-adapter/grition';
+import { Odeshin, ODESHIN_GEPP } from '../../type-script-adapter/odeshin';
+import { Plifal } from '../../type-script-adapter/plifal';
 import {
   FileAConfigurationOdeshin,
   FILE_A_CONFIGURATION_GEPP,
