+++ packages/open-schema-type-script/src/example/file/fileAConfiguration.ts
@@ -2,9 +2,9 @@ import {
   ComparisonConfigurationTypeName,
   FilePathAccessorInput,
 } from '../../utilities/file/getNestedFilePaths';
-import { Grition } from '../core/grition';
-import { Odeshin, ODESHIN_GEPP } from '../core/odeshin';
-import { Plifal } from '../core/plifal';
+import { Grition } from '../../type-script-adapter/grition';
+import { Odeshin, ODESHIN_GEPP } from '../../type-script-adapter/odeshin';
+import { Plifal } from '../../type-script-adapter/plifal';
 
 export const FILE_A_CONFIGURATION_IDENTIFIER = 'file-a-configuration' as const;
