+++ packages/open-schema-type-script/src/example/file/typeScriptFileC.ts
@@ -3,9 +3,9 @@ import { OnamaEstinant } from '../../core/estinant';
 import { TropoignantTypeName } from '../../core/tropoignant';
 import { File } from '../../utilities/file/file';
 import { FileExtensionSuffixIdentifier } from '../../utilities/file/fileExtensionSuffixIdentifier';
-import { Grition } from '../core/grition';
-import { Odeshin, ODESHIN_GEPP } from '../core/odeshin';
-import { Plifal } from '../core/plifal';
+import { Grition } from '../../type-script-adapter/grition';
+import { Odeshin, ODESHIN_GEPP } from '../../type-script-adapter/odeshin';
+import { Plifal } from '../../type-script-adapter/plifal';
 import {
   TypeScriptFileBOdeshin,
   TYPE_SCRIPT_FILE_B_GEPP,
