+++ packages/open-schema-type-script/src/type-script-adapter/grition.ts
