+++ packages/open-schema-type-script/src/type-script-adapter/plifal.ts
@@ -1,5 +1,5 @@
-import { Gepp } from '../../core/gepp';
-import { Quirm } from '../../core/quirm';
+import { Gepp } from '../core/gepp';
+import { Quirm } from '../core/quirm';
 import { Odeshin, OdeshinGepp } from './odeshin';
 
 /**
